requirements.txt
Flask==3.0.0
Flask
RPi.GPIO
requests
psutil
Flask-CORS==4.0.0
psutil==5.9.8
